horizontal decoration in resource
