+++ objects/checks/network_unreachable_response.rb
@@ -21,40 +21,15 @@
 # SOFTWARE.
 
 require 'net/http'
-require 'openssl'
-require_relative 'http_response'
-require_relative 'checked_response'
-require_relative 'checks/zlib_buffer_error_response'
-require_relative 'checks/socket_error_response'
-require_relative 'checks/timedout_response'
-require_relative 'checks/network_unreachable_response'
-require_relative 'checks/proxy_authentication_required_response'
+require_relative '../internal_error_response'
 
 #
-# Single web resource.
+# Response checked for network unreachable error
 #
-class Resource
-  def initialize(uri)
-    @uri = uri
-  end
-
-  def take(host = nil, port = nil)
-    http = Net::HTTP.new(@uri.host, @uri.port, host, port)
-    if @uri.scheme == 'https'
-      http.use_ssl = true
-      http.verify_mode = OpenSSL::SSL::VERIFY_NONE
-    end
-    req = Net::HTTP::Get.new(@uri.request_uri)
-    req['User-Agent'] = 'SixNines.io (not Firefox, Chrome, or Safari)'
-    CheckedResponse.new(
-      HTTPResponse.new(http, req),
-      [
-        ZlibBufferErrorResponse.new,
-        ProxyAuthenticationRequiredResponse.new,
-        SocketErrorResponse.new(3),
-        TimedoutResponse.new(5),
-        NetworkUnreachableResponse.new
-      ]
-    ).receive
+class NetworkUnreachableResponse
+  def check(response)
+    response.receive
+  rescue Errno::ENETUNREACH
+    InternalErrorResponse.new('Network unreachable.').receive
   end
 end
