+++ objects/http_response.rb
