+++ objects/internal_error_from_exception_response.rb
