+++ objects/internal_error_response.rb
